requirements.txt
absl-py==0.11.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037829118/work
astor==0.8.1
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
brotlipy==0.7.0
cached-property==1.5.2
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247269562/work
chardet @ file:///C:/ci/chardet_1607706912142/work
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1613401633207/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///home/ktietz/src/ci_mi/defusedxml_1612808095953/work
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
jupyterlab==2.2.6
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
jupyterlab-widgets==1.0.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclient @ file:///tmp/build/80754af9/nbclient_1613676291033/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1611348388909/work
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandocfilters @ file:///C:/ci/pandocfilters_1605120535071/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit==3.0.16
protobuf==3.15.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.8.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
scipy==1.5.4
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
